# задание/openapi.yml
openapi: "3.0.1"
info:
  title: Tender Management API
  version: "1.0"
  description: |
    API для управления тендерами и предложениями. 

    Основные функции API включают управление тендерами (создание, изменение, получение списка) и управление предложениями (создание, изменение, получение списка).
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер API
paths:
...
  /tenders/{tenderId}/status:
    get:
      summary: Получение текущего статуса тендера
      description: Получить статус тендера по его уникальному идентификатору.
      operationId: getTenderStatus
      parameters:
        - name: tenderId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/tenderId"
        - name: username
          in: query
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Текущий статус тендера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tenderStatus"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Тендер не найден.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
    put:
      summary: Изменение статуса тендера
      description: Изменить статус тендера по его идентификатору.
      operationId: updateTenderStatus
      parameters:
        - name: tenderId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/tenderId"
        - name: status
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/tenderStatus"
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Статус тендера успешно изменен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tender"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Тендер не найден.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
...
components:
  schemas:
    username:
      type: string
      description: Уникальный slug пользователя.
      example: test_user
    tenderStatus:
      type: string
      description: Статус тендер
      enum:
        - Created
        - Published
        - Closed
    tenderServiceType:
      type: string
      description: Вид услуги, к которой относиться тендер
      enum:
        - Construction
        - Delivery
        - Manufacture
    tenderId:
      type: string
      description: Уникальный идентификатор тендера, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
    tenderName:
      type: string
      description: Полное название тендера
      maxLength: 100
    tenderDescription:
      type: string
      description: Описание тендера
      maxLength: 500
    tenderVersion:
      type: integer
      description: Номер версии посел правок
      format: int32
      minimum: 1
      default: 1
    organizationId:
      type: string
      description: Уникальный идентификатор организации, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
    tender:
      type: object
      description: Информация о тендере
      properties:
        id:
          $ref: "#/components/schemas/tenderId"
        name:
          $ref: "#/components/schemas/tenderName"
        description:
          $ref: "#/components/schemas/tenderDescription"
        serviceType:
          $ref: "#/components/schemas/tenderServiceType"
        status:
          $ref: "#/components/schemas/tenderStatus"
        organizationId:
          $ref: "#/components/schemas/organizationId"
        version:
          $ref: "#/components/schemas/tenderVersion"
        createdAt:
          type: string
          description: |
            Серверная дата и время в момент, когда пользователь отправил тендер на создание.
            Передается в формате RFC3339.
          example: 2006-01-02T15:04:05Z07:00
        
      required:
        - id
        - name
        - description
        - serviceType
        - status
        - organizationId
        - version
        - createdAt
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        name: Доставка товары Казань - Москва
        description: Нужно доставить оборудовоние для олимпиады по робототехники
        status: Created
        serviceType: Delivery
        version: 1
        createdAt: 2006-01-02T15:04:05Z07:00
...
    errorResponse:
      type: object
      description: Используется для возвращения ошибки пользователю
      properties:
        reason:
          type: string
          description: Описание ошибки в свободной форме
          minLength: 5
      required:
        - reason
      example:
        reason: <объяснение, почему запрос пользователя не может быть обработан>
...